+++ tests/test_exceptions.py
@@ -18,12 +18,12 @@ def test_InvalidMessageInputError(capsys):
     THEN assert it raises exception and prints proper output
     """
     with pytest.raises(InvalidMessageInputError):
-        raise InvalidMessageInputError('Email', 'from_', 'email address')
+        raise InvalidMessageInputError('Email', 'from_', 'some_value', 'email address')
         out, err = capsys.readouterr()
-        expected = 'Invalid input for specified message class: Email'
-        expected += '\n\t* argument: "from_"'
-        expected += '\n\t* input type must be: email address\n'
-        assert out == expected
+        assert 'Invalid input for specified message class: Email' in out
+        assert '* argument: "from_"' in out
+        assert '* value given: some_value'
+        assert '* input type must be: email address' in out
         assert err == ''
 
 
@@ -36,8 +36,7 @@ def test_UnsupportedMessageTypeError_default(capsys):
     with pytest.raises(UnsupportedMessageTypeError):
         raise UnsupportedMessageTypeError('BadType')
         out, err = capsys.readouterr()
-        expected = 'Invalid input for specified message class: BadType\n'
-        assert out == expected
+        assert 'Invalid input for specified message class: BadType' in out
         assert err == ''
 
 
@@ -50,10 +49,9 @@ def test_UnsupportedMessageTypeError_listarg(capsys):
     with pytest.raises(UnsupportedMessageTypeError):
         raise UnsupportedMessageTypeError('BadType', {'m1', 'm2'})
         out, err = capsys.readouterr()
-        expected = 'Invalid input for specified message class: BadType'
-        expected += '\n\t* Supported message types: '
-        expected += "{'m1', 'm2'}\n"
-        assert out == expected
+        assert 'Invalid input for specified message class: BadType' in out
+        assert '* Supported message types:' in out
+        assert "{'m1', 'm2'}" in out
         assert err == ''
 
 
@@ -66,6 +64,5 @@ def test_UnknownProfileError(capsys):
     with pytest.raises(UnknownProfileError):
         raise UnknownProfileError('unknown_user')
         out, err = capsys.readouterr()
-        expected = 'Unknown Profile name: unknown_user\n'
-        assert out == expected
+        assert 'Unknown Profile name: unknown_user' in out
         assert err == ''