+++ tests/test_utils.py
@@ -1,11 +1,13 @@
 """messages._utils tests."""
 
 import builtins
+import re
 
 import pytest
 
 import messages._utils
-
+from messages._utils import credential_property
+from messages._utils import validate_property
 from messages._utils import validate_input
 from messages._utils import check_valid
 from messages._utils import validate_email
@@ -13,6 +15,7 @@ from messages._utils import validate_twilio
 from messages._utils import validate_slackwebhook
 from messages._utils import validate_slackpost
 from messages._utils import validate_telegrambot
+from messages._utils import timestamp
 from messages._utils import validus
 from messages._exceptions import InvalidMessageInputError
 
@@ -21,282 +24,218 @@ from messages._exceptions import InvalidMessageInputError
 # FIXTURES
 ##############################################################################
 
-class Email:
-    """Basic Email class used for testing."""
-    def __init__(self, from_, to , cc, bcc):
-        self.from_, self.to, self.cc, self.bcc = from_, to, cc, bcc
+class DummyClass:
+    """Test class used for testing."""
+    cred_param = credential_property('cred_param')
+    normal_param = validate_property('normal_param')
+    def __init__(self, cred_param, normal_param):
+        self.cred_param, self.normal_param = cred_param, normal_param
 
 
-class Twilio:
-    """Basic Twilio class used for testing."""
-    def __init__(self, from_, to , media_url):
-        self.from_, self.to, self.media_url = from_, to, media_url
+@pytest.fixture()
+def get_test_class():
+    """Return an instance of TestClass."""
+    return DummyClass('s3cr3t', 'information')
 
 
-class SlackWebhook:
-    """Basic SlackWebhook class used for testing."""
-    def __init__(self, webhook_url, attachments):
-       self.url = webhook_url
-       self.attachments = attachments
-
-
-class SlackPost:
-    """Basic SlackPost class used for testing."""
-    def __init__(self, token, channel):
-       self.credentials = token
-       self.channel = channel
-
-
-class TelegramBot:
-    """Basic TelegramBot class used for testing."""
-    def __init__(self, bot_token, chat_id):
-       self.bot_token = bot_token
-       self.chat_id = chat_id
-
-
-def func(item):
+def val_test_func(item):
     """Test func for check_valid."""
     if item == 'BAD':
         return False
     return True
 
 
-@pytest.fixture()
-def get_email():
-    """Return a valid Email object for testing."""
-    return Email('me@here.com', 'you@there.com',
-        ['him@there.com', 'her@there.com'], ['them@there.com'])
+##############################################################################
+# TEST: credential_property & validate_property
+##############################################################################
+
+def test_properties_instantiation(get_test_class, mocker):
+    """
+    GIVEN a message class that utilized both *_property factory functions
+    WHEN instantiation occurs
+    THEN assert credential_property and validate_property successfully execute
+    """
+    val_mock = mocker.patch.object(messages._utils, 'validate_input')
+    t = get_test_class
+    assert t.cred_param == '***obfuscated***'
+    assert '_cred_param' in t.__dict__
+    assert t._cred_param == 's3cr3t'
+    assert t.normal_param == 'information'
+    t.normal_param = 'new information'
+    t.normal_param = 'even newer information'
+    assert val_mock.call_count == 2
 
 
-@pytest.fixture()
-def get_twilio():
-    """Return a valid Twilio object for testing."""
-    return Twilio('+3215556666', '+3216665555', 'https://url.com')
+def test_credential_property(mocker):
+    """
+    GIVEN an attribute to make as a property
+    WHEN credential_property is called
+    THEN assert a property is returned
+    """
+    val_mock = mocker.patch.object(messages._utils, 'validate_input')
+    param = credential_property('param')
+    assert isinstance(param, property)
 
 
-@pytest.fixture()
-def get_slackwebhook():
-    """Return a valid SlackWebhook object for testing."""
-    return SlackWebhook('https://webhookurl.com', 'https://url.com')
-
-
-@pytest.fixture()
-def get_slackpost():
-    """Return a valid SlackWebhook object for testing."""
-    return SlackPost('12345abcdef', 'general')
-
-
-@pytest.fixture()
-def get_tgram():
-    """Return a valid SlackWebhook object for testing."""
-    return TelegramBot('12345:ABCDEFG', '12356')
+def test_validate_property(mocker):
+    """
+    GIVEN an attribute to make as a property
+    WHEN validate_property is called
+    THEN assert a property is returned
+    """
+    val_mock = mocker.patch.object(messages._utils, 'validate_input')
+    param = validate_property('param')
+    assert isinstance(param, property)
 
 
 ##############################################################################
 # TEST: validate_input
 ##############################################################################
 
-def test_val_input_NotSupported(get_email, mocker):
+def test_val_input_NotSupported():
     """
     GIVEN a message object is instantiated
     WHEN validate_input() is called on a message object that is
         not supported for input validation
     THEN assert nothing happens
     """
-    class Nothing:
-        def __init__(self, name):
-            self.name = name
-
-    n = Nothing('nothing')
-    validate_input(n, 'name')
+    result = validate_input('NotSupported', 'attr', 'some_value')
+    assert result == 1
 
 
-def test_val_input_Email(get_email, mocker):
+@pytest.mark.parametrize('msg_type, func, result', [
+    ('Email', 'validate_email', 0),
+    ('Twilio', 'validate_twilio', 0),
+    ('SlackPost', 'validate_slackpost', 0),
+    ('SlackWebhook', 'validate_slackwebhook', 0),
+    ('TelegramBot', 'validate_telegrambot', 0),
+])
+def test_val_input_supported(msg_type, func, result, mocker):
     """
     GIVEN a message object is instantiated
     WHEN validate_input() is called on a message object
-    THEN assert the proper valid_* functions are called
+    THEN assert the proper validate_* func is called and 0 is the return value
     """
-    val_mock = mocker.patch.object(messages._utils, 'validate_email')
-    e = get_email
-    for key in e.__dict__.keys():
-        validate_input(e, key)
-    assert val_mock.call_count == 4
-
-
-def test_val_input_Twilio(get_twilio, mocker):
-    """
-    GIVEN a message object is instantiated
-    WHEN validate_input() is called on a message object
-    THEN assert the proper valid_* functions are called
-    """
-    val_mock = mocker.patch.object(messages._utils, 'validate_twilio')
-    e = get_twilio
-    for key in e.__dict__.keys():
-        validate_input(e, key)
-    assert val_mock.call_count == 3
-
-
-def test_val_input_SlackWebhook(get_slackwebhook, mocker):
-    """
-    GIVEN a message object is instantiated
-    WHEN validate_input() is called on a message object
-    THEN assert the proper valid_* functions are called
-    """
-    val_mock = mocker.patch.object(messages._utils, 'validate_slackwebhook')
-    e = get_slackwebhook
-    for key in e.__dict__.keys():
-        validate_input(e, key)
-    assert val_mock.call_count == 2
-
-
-def test_val_input_SlackPost(get_slackpost):
-    """
-    GIVEN a message object is instantiated
-    WHEN validate_input() is called on a message object
-    THEN assert no errors occur
-    """
-    e = get_slackpost
-    for key in e.__dict__.keys():
-        validate_input(e, key)
-
-
-def test_val_input_SlackPost_raises(get_slackpost):
-    """
-    GIVEN a message object is instantiated with bad inputs
-    WHEN validate_input() is called on a message object
-    THEN assert InvalidMessageInputError is raised
-    """
-    e = get_slackpost
-    e.credentials = 12345
-    with pytest.raises(InvalidMessageInputError):
-        for key in e.__dict__.keys():
-            validate_input(e, key)
-
-
-def test_val_input_TelegramBot(get_tgram, mocker):
-    """
-    GIVEN a message object is instantiated
-    WHEN validate_input() is called on a message object
-    THEN assert the proper valid_* functions are called
-    """
-    val_mock = mocker.patch.object(messages._utils, 'validate_telegrambot')
-    e = get_tgram
-    for key in e.__dict__.keys():
-        validate_input(e, key)
-    assert val_mock.call_count == 2
+    val_mock = mocker.patch.object(messages._utils, func)
+    r = validate_input(msg_type, 'attr', 'some_value')
+    assert val_mock.call_count == 1
+    assert r == result
 
 
 ##############################################################################
 # TEST: validate_*
 ##############################################################################
 
-def test_val_email(get_email, mocker):
+@pytest.mark.parametrize('msg_type, func, attr', [
+    ('Email', validate_email, 'address'),
+    ('Twilio', validate_twilio, 'from_'),
+    ('Twilio', validate_twilio, 'media_url'),
+    ('SlackWebhook', validate_slackwebhook, 'url'),
+    ('TelegramBot', validate_telegrambot, 'id_num' )
+
+])
+def test_val_funcs(msg_type, func, attr, mocker):
     """
-    GIVEN an Email object
-    WHEN validate_email is called
+    GIVEN a message instance
+    WHEN validate_* is called
     THEN assert check_valid is called the requisite number of times
     """
     check_mock = mocker.patch.object(messages._utils, 'check_valid')
-    e = get_email
-    e.not_checked = 'this attr should not get checked'
-    for key in e.__dict__.keys():
-        validate_email(e, key)
-    assert check_mock.call_count == 4
-
-
-def test_val_twilio(get_twilio, mocker):
-    """
-    GIVEN a Twilio object
-    WHEN validate_twilio is called
-    THEN assert check_valid is called the requisite number of times
-    """
-    check_mock = mocker.patch.object(messages._utils, 'check_valid')
-    e = get_twilio
-    e.not_checked = 'this attr should not get checked'
-    for key in e.__dict__.keys():
-        validate_twilio(e, key)
-    assert check_mock.call_count == 3
-
-
-def test_val_slackwebhook(get_slackwebhook, mocker):
-    """
-    GIVEN a SlackWebhook object
-    WHEN validate_slackwebhook is called
-    THEN assert check_valid is called the requisite number of times
-    """
-    check_mock = mocker.patch.object(messages._utils, 'check_valid')
-    e = get_slackwebhook
-    e.not_checked = 'this attr should not get checked'
-    for key in e.__dict__.keys():
-        validate_slackwebhook(e, key)
-    assert check_mock.call_count == 2
-
-
-def test_val_slackpost(get_slackpost, mocker):
-    """
-    GIVEN a SlackPost object
-    WHEN validate_slackpost is called
-    THEN assert check_valid is called the requisite number of times
-    """
-    check_mock = mocker.patch.object(messages._utils, 'check_valid')
-    e = get_slackpost
-    e.not_checked = 'this attr should not get checked'
-    for key in e.__dict__.keys():
-        validate_slackpost(e, key)
-    assert check_mock.call_count == 0
-
-
-def test_val_telegrambot(get_tgram, mocker):
-    """
-    GIVEN a TelegramBot object
-    WHEN validate_telegrambot is called
-    THEN assert check_valid is called the requisite number of times
-    """
-    check_mock = mocker.patch.object(messages._utils, 'check_valid')
-    e = get_tgram
-    e.not_checked = 'this attr should not get checked'
-    for key in e.__dict__.keys():
-        validate_telegrambot(e, key)
+    func(attr, 'some_value')
     assert check_mock.call_count == 1
+
+
+def test_val_slackPost(mocker):
+    """
+    GIVEN a slackPost instance
+    WHEN validate_slackpost is called with the correct input type
+    THEN assert nothing bad happens
+    **SlackPost is tested separately from the parametrized test_val_func()
+    test because validate_slackpost is written differently and cannot be
+    tested the same way.
+    """
+    check_mock = mocker.patch.object(messages._utils, 'check_valid')
+    validate_slackpost('channel', 'some_value')
+    validate_slackpost('attachments', 'some_value')
+
+
+def test_val_slackPost_raises():
+    """
+    GIVEN a slackPost instance
+    WHEN validate_slackpost is called with incorrect input type
+    THEN assert InvalideMessageInputError is raised
+    **SlackPost is tested separately from the parametrized test_val_func()
+    test because validate_slackpost is written differently and cannot be
+    tested the same way.
+    """
+    with pytest.raises(InvalidMessageInputError):
+        validate_slackpost('channel', 1)
 
 
 ##############################################################################
 # TEST: check_valid
 ##############################################################################
 
-def test_check_valid(get_email):
+def test_check_valid(get_test_class):
     """
     GIVEN a valid message object
     WHEN check_valid is called on requisite attributes
     THEN assert normal behavior and no exceptions raised
     """
-    e = get_email
-    for key in e.__dict__.keys():
-        check_valid(e, key, func, 'email')
+    t = get_test_class
+    for key, value in t.__dict__.items():
+        check_valid('TestClass', key, value, val_test_func, 'required type')
 
 
-def test_check_valid_singleton_raisesExc(get_email):
+def test_check_valid_listAttributes(get_test_class):
+    """
+    GIVEN a valid message object
+    WHEN check_valid is called on requisite attributes that are lists
+    THEN assert normal behavior and no exceptions raised
+    """
+    t = get_test_class
+    t.listAttr = ['val1', 'val2']
+    for key, value in t.__dict__.items():
+        check_valid('TestClass', key, value, val_test_func, 'required type')
+
+
+def test_check_valid_singleton_raisesExc(get_test_class):
     """
     GIVEN a message object with a single invalid input
     WHEN check_valid is called
     THEN assert InvalidMessageInputError is raised
     """
-    e = get_email
-    e.from_ = 'BAD'
+    t = get_test_class
+    t.normal_param = 'BAD'
     with pytest.raises(InvalidMessageInputError):
-        for key in e.__dict__.keys():
-            check_valid(e, key, func, 'email')
+        for key, value in t.__dict__.items():
+            check_valid('TestClass', key, value, val_test_func, 'required type')
 
 
-def test_check_valid_list_raisesExc(get_email):
+def test_check_valid_list_raisesExc(get_test_class):
     """
     GIVEN a message object with a list of invalid inputs
     WHEN check_valid is called
     THEN assert InvalidMessageInputError is raised
     """
-    e = get_email
-    e.to = ['BAD', 'BAD']
+    t = get_test_class
+    t.normal_param = ['GOOD', 'BAD']
     with pytest.raises(InvalidMessageInputError):
-        for key in e.__dict__.keys():
-            check_valid(e, key, func, 'email')
+        for key, value in t.__dict__.items():
+            check_valid('TestClass', key, value, val_test_func, 'required_type')
+
+
+##############################################################################
+# TEST: timestamp
+##############################################################################
+
+def test_timestamp():
+    """
+    GIVEN a need for a timestamp (i.e. --verbose output)
+    WHEN timestamp() is called
+    THEN assert a timestamp as a string is returned
+    """
+    t = timestamp()
+    #r = '^\d{4}-[A-Za-z]{3}-\d{1,2}\s{1}\d{2}:\d{2}:\d{2}$'
+    #assert re.match(r, t)
+    assert isinstance(t, str)